+++ PsIdoIt/psidoit.psm1
@@ -19,5 +19,5 @@ ForEach($ImportScriptFile In @($Public + $Private) )
 # Read in or create an initial config file and variable
 # Export Public functions ($Public.BaseName) for WIP modules
 # Set variables visible to the module and its functions only
-
+# Write-Host $Public.Basename
 Export-ModuleMember -Function $Public.Basename